refactor, separate by dirs
