+++ .github/scripts/js/e2e/report/messenger-report.js
@@ -12,384 +12,21 @@
 
 const fs = require("fs");
 
-const { listMatchingFiles } = require("./fs-utils");
-
-const genericArtifactMissingLabel = "E2E REPORT ARTIFACT NOT FOUND";
-
-/**
- * Builds a user-facing status line for a cluster row or fallback report.
- *
- * @param {string} status Normalized cluster status.
- * @param {string} stageLabel Human-readable stage label.
- * @returns {string} Rendered status message.
- */
-function buildStatusMessage(status, stageLabel) {
-  if (status === "cancelled") {
-    return `⚠️ ${stageLabel} CANCELLED`;
-  }
-
-  if (status === "failure") {
-    return `❌ ${stageLabel} FAILED`;
-  }
-
-  if (status === "missing") {
-    return `⚠️ ${stageLabel}`;
-  }
-
-  if (status === "success") {
-    return "✅ SUCCESS";
-  }
-
-  return stageLabel;
-}
-
-/**
- * Creates a synthetic cluster report when the expected JSON artifact is absent.
- *
- * This allows the final messenger message to stay informative even when the
- * report-preparation step failed or never produced an artifact.
- *
- * @param {string} clusterName Cluster or storage name.
- * @returns {Record<string, any>} Synthetic report payload.
- */
-function createMissingReport(clusterName) {
-  return {
-    schemaVersion: 1,
-    cluster: clusterName,
-    storageType: clusterName,
-    reportKind: "artifact-missing",
-    status: "missing",
-    statusMessage: buildStatusMessage("missing", genericArtifactMissingLabel),
-    failedStage: "artifact-missing",
-    failedStageLabel: genericArtifactMissingLabel,
-    branch: "",
-    workflowRunUrl: "",
-    clusterStatus: {
-      status: "missing",
-      stage: "artifact-missing",
-      stageLabel: genericArtifactMissingLabel,
-      message: buildStatusMessage("missing", genericArtifactMissingLabel),
-      reason: "cluster-report-artifact-missing",
-    },
-    testStatus: {
-      status: "not-run",
-      reason: "cluster-report-artifact-missing",
-      message: "E2E status is unavailable because cluster report artifact was not found",
-    },
-    metrics: {
-      passed: 0,
-      failed: 0,
-      errors: 0,
-      skipped: 0,
-      total: 0,
-      successRate: 0,
-    },
-    failedTests: [],
-    reportSource: "missing-artifact",
-  };
-}
-
-/**
- * Escapes markdown table cell content and normalizes whitespace.
- *
- * @param {any} value Raw cell value.
- * @returns {string} Sanitized table cell string.
- */
-function sanitizeCell(value) {
-  return String(value || "—")
-    .replace(/\|/g, "\\|")
-    .replace(/\r?\n/g, " ")
-    .trim();
-}
-
-/**
- * Normalizes markdown list item content to a single trimmed line.
- *
- * @param {any} value Raw list item value.
- * @returns {string} Sanitized list item string.
- */
-function sanitizeListItem(value) {
-  return String(value || "")
-    .replace(/\r?\n/g, " ")
-    .trim();
-}
-
-/**
- * Formats a numeric success rate as a percentage string.
- *
- * @param {number|string} value Raw rate value.
- * @returns {string} Formatted percentage.
- */
-function formatRate(value) {
-  const rate = Number(value || 0);
-  return `${Number.isFinite(rate) ? rate.toFixed(2) : "0.00"}%`;
-}
-
-/**
- * Picks a report date from the first report that exposes `startedAt`.
- *
- * @param {Record<string, any>[]} reports Available cluster reports.
- * @returns {string} ISO date string (`YYYY-MM-DD`).
- */
-function getReportDate(reports) {
-  const datedReport = reports.find((report) => report.startedAt);
-  if (!datedReport) {
-    return new Date().toISOString().slice(0, 10);
-  }
-
-  return String(datedReport.startedAt).slice(0, 10);
-}
-
-/**
- * Orders reports by the configured cluster order and then by cluster name.
- *
- * @param {Record<string, any>[]} reports Reports to sort.
- * @param {string[]} preferredOrder Configured cluster order.
- * @returns {Record<string, any>[]} Sorted reports copy.
- */
-function sortReports(reports, preferredOrder) {
-  const orderMap = new Map(preferredOrder.map((name, index) => [name, index]));
-
-  return [...reports].sort((left, right) => {
-    const leftKey = left.storageType || left.cluster;
-    const rightKey = right.storageType || right.cluster;
-    const leftOrder = orderMap.has(leftKey)
-      ? orderMap.get(leftKey)
-      : Number.MAX_SAFE_INTEGER;
-    const rightOrder = orderMap.has(rightKey)
-      ? orderMap.get(rightKey)
-      : Number.MAX_SAFE_INTEGER;
-
-    if (leftOrder !== rightOrder) {
-      return leftOrder - rightOrder;
-    }
-
-    return String(left.cluster || left.storageType).localeCompare(
-      String(right.cluster || right.storageType)
-    );
-  });
-}
-
-/**
- * Renders a cluster name as a markdown link when a workflow URL is available.
- *
- * @param {Record<string, any>} report Cluster report payload.
- * @returns {string} Markdown link or plain sanitized cluster name.
- */
-function formatClusterLink(report) {
-  const clusterName = sanitizeCell(report.cluster || report.storageType);
-  return report.workflowRunUrl
-    ? `[${clusterName}](${report.workflowRunUrl})`
-    : clusterName;
-}
-
-/**
- * Extracts the normalized cluster key from a report payload.
- *
- * @param {Record<string, any>} report Cluster report payload.
- * @returns {string} Cluster key or an empty string when it is missing.
- */
-function getReportClusterKey(report) {
-  return String(report.storageType || report.cluster || "").trim();
-}
-
-/**
- * Tells whether the report represents a missing artifact rather than a real
- * cluster-stage failure.
- *
- * @param {Record<string, any>} report Cluster report payload.
- * @returns {boolean} True when the report describes a missing artifact.
- */
-function isMissingReport(report) {
-  return (
-    (report.testStatus && report.testStatus.status === "missing") ||
-    (report.clusterStatus && report.clusterStatus.status === "missing") ||
-    report.reportKind === "artifact-missing" ||
-    report.failedStage === "artifact-missing" ||
-    report.status === "missing"
-  );
-}
-
-/**
- * Tells whether the report describes a failed cluster setup stage.
- *
- * @param {Record<string, any>} report Cluster report payload.
- * @returns {boolean} True for cluster-stage failures.
- */
-function isClusterFailureReport(report) {
-  if (report.clusterStatus) {
-    return (
-      report.clusterStatus.status !== "success" &&
-      report.clusterStatus.status !== "missing"
-    );
-  }
-
-  return report.reportKind !== "tests" && !isMissingReport(report);
-}
-
-/**
- * Tells whether the report should be rendered in the E2E test results table.
- *
- * @param {Record<string, any>} report Cluster report payload.
- * @returns {boolean} True for reports with test status data.
- */
-function isTestResultReport(report) {
-  if (report.clusterStatus && report.clusterStatus.status !== "success") {
-    return false;
-  }
-
-  if (report.testStatus) {
-    return (
-      report.testStatus.status !== "not-run" &&
-      report.testStatus.status !== "missing"
-    );
-  }
-
-  return report.reportKind === "tests";
-}
-
-/**
- * Normalizes the configured Loop API base URL to the `/api/v4/posts` endpoint.
- *
- * @param {string} value Raw Loop API base URL.
- * @returns {string} Normalized posts endpoint URL or an empty string.
- */
-function normalizeLoopApiBaseUrl(value) {
-  const trimmedValue = String(value || "")
-    .trim()
-    .replace(/\/+$/, "");
-
-  if (!trimmedValue) {
-    return "";
-  }
-
-  if (trimmedValue.endsWith("/api/v4/posts")) {
-    return trimmedValue;
-  }
-
-  if (trimmedValue.endsWith("/api/v4")) {
-    return `${trimmedValue}/posts`;
-  }
-
-  return `${trimmedValue}/api/v4/posts`;
-}
-
-/**
- * Reads and normalizes the Loop posts API URL from environment variables.
- *
- * @param {NodeJS.ProcessEnv} [env=process.env] Environment variables source.
- * @returns {string} Normalized posts endpoint URL or an empty string.
- */
-function getLoopPostsApiUrl(env = process.env) {
-  return normalizeLoopApiBaseUrl(env.LOOP_API_BASE_URL);
-}
-
-/**
- * Parses the configured cluster list passed via workflow environment variables.
- *
- * @param {string} value JSON-encoded cluster list.
- * @returns {string[]} Ordered cluster names.
- */
-function parseConfiguredClusters(value) {
-  const parsedValue = JSON.parse(value || "[]");
-  return Array.isArray(parsedValue) ? parsedValue : [];
-}
-
-/**
- * Reads messenger configuration from the environment prepared by the workflow.
- *
- * @param {NodeJS.ProcessEnv} [env=process.env] Environment variables source.
- * @returns {{
- *   reportsDir: string,
- *   configuredClusters: string[],
- *   loop: {
- *     apiUrl: string,
- *     channelId: string,
- *     token: string
- *   }
- * }} Normalized messenger configuration.
- */
-function readMessengerConfigFromEnv(env = process.env) {
-  const configuredClusters = parseConfiguredClusters(env.STORAGE_TYPES);
-
-  return {
-    reportsDir: env.REPORTS_DIR || "downloaded-artifacts",
-    configuredClusters,
-    loop: {
-      apiUrl: getLoopPostsApiUrl(env),
-      channelId: String(env.LOOP_CHANNEL_ID || "").trim(),
-      token: String(env.LOOP_TOKEN || "").trim(),
-    },
-  };
-}
-
-/**
- * Parses a Loop API response body if it is JSON, otherwise returns an empty
- * object and emits a warning for diagnostics.
- *
- * @param {string} responseText Raw response body.
- * @param {{ warning(message: string): void }} core GitHub core API.
- * @returns {Record<string, any>} Parsed response payload or an empty object.
- */
-function parseLoopApiPayload(responseText, core) {
-  if (!responseText) {
-    return {};
-  }
-
-  try {
-    return JSON.parse(responseText);
-  } catch (error) {
-    core.warning(
-      `Loop API returned a non-JSON response body: ${error.message}`
-    );
-    return {};
-  }
-}
-
-/**
- * Sends a single post to Loop and returns the parsed API payload.
- *
- * @param {{
- *   apiUrl: string,
- *   channelId: string,
- *   token: string,
- *   message: string,
- *   rootId?: string
- * }} request Loop API request payload.
- * @param {{
- *   info(message: string): void,
- *   warning(message: string): void
- * }} core GitHub core API.
- * @returns {Promise<Record<string, any>>} Parsed Loop API response.
- */
-async function postToLoopApi(
-  { apiUrl, channelId, token, message, rootId },
-  core
-) {
-  const response = await fetch(apiUrl, {
-    method: "POST",
-    headers: {
-      Authorization: `Bearer ${token}`,
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      channel_id: channelId,
-      message,
-      ...(rootId ? { root_id: rootId } : {}),
-    }),
-  });
-  const responseText = await response.text();
-
-  if (!response.ok) {
-    throw new Error(
-      `Loop API request failed with status ${response.status}: ${responseText}`
-    );
-  }
-
-  const payload = parseLoopApiPayload(responseText, core);
-  core.info(`Loop API accepted report with status ${response.status}`);
-  return payload;
-}
+const { listMatchingFiles } = require("./shared/fs-utils");
+const { publishToLoop } = require("./messenger/loop-client");
+const {
+  getLoopPostsApiUrl,
+  readMessengerConfigFromEnv,
+} = require("./messenger/config");
+const {
+  createMissingReport,
+  getReportClusterKey,
+  sortReports,
+} = require("./messenger/model");
+const {
+  buildMainMessage,
+  buildThreadMessages,
+} = require("./messenger/markdown");
 
 /**
  * Loads report JSON files from disk and injects synthetic reports for clusters
@@ -431,216 +68,7 @@ function readReports(reportsDir, configuredClusters, core) {
     }
   }
 
-  const orderedReports = sortReports(
-    Array.from(reportsByCluster.values()),
-    configuredClusters
-  );
-  return orderedReports;
-}
-
-/**
- * Renders the top-level messenger markdown message.
- *
- * @param {Record<string, any>[]} orderedReports Reports ordered for display.
- * @returns {string} Main markdown message.
- */
-function buildMainMessage(orderedReports) {
-  const reportDate = getReportDate(orderedReports);
-  const branches = Array.from(
-    new Set(orderedReports.map((report) => report.branch).filter(Boolean))
-  );
-  const lines = [`## DVP | E2E on nested clusters | ${reportDate}`, ""];
-
-  if (branches.length === 1 && branches[0] !== "main") {
-    lines.push(`Branch: \`${branches[0]}\``);
-    lines.push("");
-  }
-
-  const testsReports = orderedReports.filter(
-    (report) => isTestResultReport(report) && getReportClusterKey(report)
-  );
-  const stageFailureReports = orderedReports.filter(
-    (report) => isClusterFailureReport(report) && getReportClusterKey(report)
-  );
-  const missingReports = orderedReports.filter(
-    (report) =>
-      isMissingReport(report) &&
-      !isClusterFailureReport(report) &&
-      getReportClusterKey(report)
-  );
-
-  if (testsReports.length > 0) {
-    lines.push("### Test results");
-    lines.push("");
-    lines.push(
-      "| Cluster | ✅ Passed | ⏭️ Skipped | ❌ Failed | ⚠️ Errors | Total | Success Rate |"
-    );
-    lines.push("|---|---:|---:|---:|---:|---:|---:|");
-
-    for (const report of testsReports) {
-      const metrics = report.metrics || {};
-      lines.push(
-        `| ${formatClusterLink(report)} | ${metrics.passed || 0} | ${
-          metrics.skipped || 0
-        } | ${metrics.failed || 0} | ${metrics.errors || 0} | ${
-          metrics.total || 0
-        } | ${formatRate(metrics.successRate)} |`
-      );
-    }
-
-    lines.push("");
-  }
-
-  if (stageFailureReports.length > 0) {
-    lines.push("### Cluster failures");
-    lines.push("");
-
-    for (const report of stageFailureReports) {
-      lines.push(
-        `- ${formatClusterLink(report)}: ${sanitizeListItem(
-          (report.clusterStatus && report.clusterStatus.message) ||
-            report.statusMessage ||
-            report.failedStageLabel ||
-            report.failedStage
-        )}`
-      );
-    }
-
-    lines.push("");
-  }
-
-  if (missingReports.length > 0) {
-    lines.push("### Missing reports");
-    lines.push("");
-
-    for (const report of missingReports) {
-      const missingMessage =
-        report.clusterStatus && report.clusterStatus.status === "missing"
-          ? report.clusterStatus.message
-          : report.testStatus && report.testStatus.message;
-      lines.push(
-        `- ${formatClusterLink(report)}: ${sanitizeListItem(
-          missingMessage ||
-            (report.clusterStatus && report.clusterStatus.message) ||
-            report.statusMessage ||
-            report.failedStageLabel ||
-            report.failedStage
-        )}`
-      );
-    }
-
-    lines.push("");
-  }
-
-  return lines.join("\n").trim();
-}
-
-/**
- * Tells whether the report should contribute failed-test details to the thread.
- *
- * @param {Record<string, any>} report Cluster report payload.
- * @returns {boolean} True when failed-test details should be rendered.
- */
-function hasFailedTests(report) {
-  if (Array.isArray(report.failedTests) && report.failedTests.length > 0) {
-    return true;
-  }
-
-  return Boolean(
-    report.testStatus &&
-      (report.testStatus.status === "failure" ||
-        report.testStatus.status === "cancelled") ||
-    (report.metrics && report.metrics.failed) ||
-      (report.metrics && report.metrics.errors)
-  );
-}
-
-/**
- * Extracts the top-level test group name from a failed test title.
- *
- * For Ginkgo titles like `[It] VirtualMachineOperationRestore restores ...`,
- * this returns `VirtualMachineOperationRestore`.
- *
- * @param {string} testName Full failed test name.
- * @returns {string} Top-level test group label.
- */
-function getFailedTestGroupName(testName) {
-  const normalizedName = sanitizeListItem(testName).replace(/^\[[^\]]+\]\s*/, "");
-  const [groupName] = normalizedName.split(/\s+/, 1);
-  return groupName || "Unknown";
-}
-
-/**
- * Aggregates failed test names into an ordered unique group list.
- *
- * @param {string[]} failedTests Failed testcase names.
- * @returns {string[]} Ordered unique group names.
- */
-function summarizeFailedTestGroups(failedTests) {
-  const groupNames = [];
-
-  for (const testName of failedTests) {
-    const groupName = getFailedTestGroupName(testName);
-    if (!groupNames.includes(groupName)) {
-      groupNames.push(groupName);
-    }
-  }
-
-  return groupNames;
-}
-
-/**
- * Builds the thread reply body for a single cluster with failed tests.
- *
- * @param {Record<string, any>} report Cluster report payload.
- * @returns {string} Cluster-specific failed tests markdown.
- */
-function buildFailedTestsClusterMessage(report) {
-  const clusterName = sanitizeListItem(report.cluster || report.storageType);
-  const lines = [`**${clusterName}**`];
-
-  if (Array.isArray(report.failedTests) && report.failedTests.length > 0) {
-    const failedGroups = summarizeFailedTestGroups(report.failedTests);
-    lines.push("");
-    lines.push("| Test group |");
-    lines.push("|---|");
-    for (const groupName of failedGroups) {
-      lines.push(`| ${sanitizeCell(groupName)} |`);
-    }
-  } else {
-    lines.push(
-      `- ${
-        sanitizeListItem(report.testStatus && report.testStatus.message) ||
-        "No testcase-level failures were collected, but the E2E stage reported failures."
-      }`
-    );
-  }
-
-  return lines.join("\n");
-}
-
-/**
- * Renders thread markdown messages containing failed test names, if any.
- *
- * @param {Record<string, any>[]} orderedReports Reports ordered for display.
- * @returns {string[]} Thread markdown messages in publish order.
- */
-function buildThreadMessages(orderedReports) {
-  const testsReports = orderedReports.filter(
-    (report) => isTestResultReport(report)
-  );
-  const failedTestReports = testsReports.filter(hasFailedTests);
-
-  if (failedTestReports.length === 0) {
-    return [];
-  }
-
-  return failedTestReports.map((report, index) => {
-    const clusterMessage = buildFailedTestsClusterMessage(report);
-    return index === 0
-      ? ["### Failed tests", clusterMessage].join("\n\n")
-      : clusterMessage;
-  });
+  return sortReports(Array.from(reportsByCluster.values()), configuredClusters);
 }
 
 /**
@@ -673,67 +101,6 @@ function buildMessengerMessages({
     threadMessage: threadMessages.join("\n\n"),
     threadMessages,
   };
-}
-
-/**
- * Publishes the main report and optional failed-tests thread to Loop.
- *
- * @param {{
- *   message: string,
- *   threadMessages: string[],
- *   loop: {
- *     apiUrl: string,
- *     channelId: string,
- *     token: string
- *   }
- * }} params Message payload and Loop credentials.
- * @param {{
- *   setOutput(name: string, value: string): void,
- *   info(message: string): void,
- *   warning(message: string): void
- * }} core GitHub core API.
- * @returns {Promise<void>}
- */
-async function publishToLoop({ message, threadMessages, loop }, core) {
-  if (!loop.apiUrl && !loop.channelId && !loop.token) {
-    return;
-  }
-
-  if (!loop.apiUrl || !loop.channelId || !loop.token) {
-    throw new Error(
-      "LOOP_CHANNEL_ID, LOOP_TOKEN, and LOOP_API_BASE_URL are required"
-    );
-  }
-
-  const rootPost = await postToLoopApi(
-    {
-      apiUrl: loop.apiUrl,
-      channelId: loop.channelId,
-      token: loop.token,
-      message,
-    },
-    core
-  );
-
-  let lastReplyPost = null;
-  for (const replyMessage of threadMessages) {
-    lastReplyPost = await postToLoopApi(
-      {
-        apiUrl: loop.apiUrl,
-        channelId: loop.channelId,
-        token: loop.token,
-        message: replyMessage,
-        rootId: rootPost.id,
-      },
-      core
-    );
-  }
-
-  core.setOutput("root_post_id", rootPost.id || "");
-  core.setOutput(
-    "thread_post_id",
-    lastReplyPost && lastReplyPost.id ? lastReplyPost.id : ""
-  );
 }
 
 /**
